+++ LanceOfDestiny-FourthDimension/src/main/java/org/LanceOfDestiny/domain/physics/Vector.java
@@ -1,32 +1,32 @@
 package org.LanceOfDestiny.domain.physics;
 
 public class Vector {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public Vector(float x, float y) {
+    public Vector(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 
     public Vector getNormalized() {
-        float length = (float) Math.sqrt(getX() * getX() + getY() * getY());
+        double length = (double) Math.sqrt(getX() * getX() + getY() * getY());
         if (length != 0) {
             return new Vector(getX() / length, getY() / length);
         }
@@ -34,8 +34,8 @@ public class Vector {
     }
 
 
-    public float distanceTo(Vector pos2) {
-        return (float) Math.sqrt(Math.pow(getX() - pos2.getX(), 2) + Math.pow(getY() - pos2.getY(), 2));
+    public double distanceTo(Vector pos2) {
+        return (double) Math.sqrt(Math.pow(getX() - pos2.getX(), 2) + Math.pow(getY() - pos2.getY(), 2));
     }
 
     public Vector add(Vector pos2) {
@@ -46,7 +46,7 @@ public class Vector {
         return new Vector(getX() - pos2.getX(), getY() - pos2.getY());
     }
 
-    public Vector scale(float ratio) {
+    public Vector scale(double ratio) {
         return new Vector(getX() * ratio, getY() * ratio);
     }
 
@@ -54,7 +54,7 @@ public class Vector {
         return new Vector(0,0);
     }
 
-    public float dotProduct(Vector v){
+    public double dotProduct(Vector v){
         return getX() * v.getX() + getY()* v.getY();
     }
 }
